# Improve performance and memory consumption