I am looking at integrating this gem into a legacy application on Ruby 1.8.7 and Ruby 1.9.3. The default implementation here is already ~3x faster out of the box than the legacy applications method, but we have some endpoints where 50% of the request time in spent in adding business days. By reducing the memory usage, this reduces pressure on GC as well as bypassing the cost of the initial allocations.

Each commit was tested independently, but below we can see the results for everything together.

### Before
```
Calculating -------------------------------------
             day + 1      2.966M (± 9.1%) i/s -      5.918M in   2.017065s
add_business_day(0) on weekend
                         18.890k (± 7.1%) i/s -     37.842k in   2.015050s
add_business_day(0) on weekday
                         18.451k (±14.3%) i/s -     37.240k in   2.077531s
add_business_day(0) on holiday
                          6.632k (± 9.1%) i/s -     13.902k in   2.118830s
add_business_day(0) on weekend with monday holiday
                          6.619k (± 4.6%) i/s -     13.680k in   2.071195s
add_business_day(1) on weekend
                          9.588k (± 3.7%) i/s -     20.034k in   2.092520s
add_business_day(1) on weekday
                          9.625k (± 7.2%) i/s -     19.580k in   2.046498s
add_business_day(1) on holiday
                          4.430k (± 9.6%) i/s -      9.366k in   2.140654s
add_business_day(1) on weekend with monday holiday
                          6.489k (±11.0%) i/s -     12.980k in   2.038185s
add_business_day(10) on weekend
                          1.744k (± 4.6%) i/s -      3.654k in   2.100506s
add_business_day(10) on weekday
                          1.649k (±11.9%) i/s -      3.268k in   2.024484s
add_business_day(10) on holiday
                          1.072k (±11.7%) i/s -      2.180k in   2.069246s
add_business_day(10) on weekend with monday holiday
                          1.096k (± 7.5%) i/s -      2.240k in   2.058505s
Calculating -------------------------------------
             day + 1    72.000  memsize (     0.000  retained)
                         1.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(0) on weekend
                         6.094k memsize (     0.000  retained)
                        87.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(0) on weekday
                         6.022k memsize (     0.000  retained)
                        86.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(0) on holiday
                        20.414k memsize (     0.000  retained)
                       261.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(0) on weekend with monday holiday
                        20.486k memsize (     0.000  retained)
                       262.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(1) on weekend
                        12.824k memsize (     0.000  retained)
                       173.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(1) on weekday
                        12.752k memsize (     0.000  retained)
                       172.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(1) on holiday
                        30.939k memsize (     0.000  retained)
                       393.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(1) on weekend with monday holiday
                        20.486k memsize (     0.000  retained)
                       262.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(10) on weekend
                        73.682k memsize (     0.000  retained)
                       951.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(10) on weekday
                        73.610k memsize (     0.000  retained)
                       950.000  objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(10) on holiday
                       125.952k memsize (     0.000  retained)
                         1.585k objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
add_business_day(10) on weekend with monday holiday
                       126.024k memsize (     0.000  retained)
                         1.586k objects (     0.000  retained)
                        16.000  strings (     0.000  retained)
```

### After
```
Calculating -------------------------------------
             day + 1      2.995M (± 6.1%) i/s -      6.093M in   2.042448s
is_business_day? on weekend
                          1.371M (± 3.9%) i/s -      2.778M in   2.029435s
is_business_day? on weekday
                          1.388M (± 4.5%) i/s -      2.876M in   2.075842s
is_business_day? on holiday
                          1.361M (± 5.3%) i/s -      2.723M in   2.005941s
add_business_day(0) on weekend
                        952.681k (± 3.8%) i/s -      1.956M in   2.055616s
add_business_day(0) on weekday
                        946.133k (± 3.0%) i/s -      1.932M in   2.043956s
add_business_day(0) on holiday
                        444.823k (± 3.3%) i/s -    889.790k in   2.002453s
add_business_day(0) on weekend with monday holiday
                        356.299k (± 4.4%) i/s -    716.738k in   2.015715s
add_business_day(1) on weekend
                        380.591k (± 4.3%) i/s -    776.666k in   2.044680s
add_business_day(1) on weekday
                        387.654k (± 3.5%) i/s -    797.663k in   2.060292s
add_business_day(1) on holiday
                        253.445k (± 4.6%) i/s -    523.299k in   2.069200s
add_business_day(1) on weekend with monday holiday
                        224.281k (± 4.2%) i/s -    460.257k in   2.055774s
add_business_day(10) on weekend
                         59.080k (± 8.0%) i/s -    118.140k in   2.014506s
add_business_day(10) on weekday
                         59.682k (± 5.9%) i/s -    119.760k in   2.014163s
add_business_day(10) on holiday
                         56.015k (± 5.2%) i/s -    112.580k in   2.015369s
add_business_day(10) on weekend with monday holiday
                         54.051k (± 9.1%) i/s -    113.841k in   2.126502s
Calculating -------------------------------------
             day + 1    72.000  memsize (     0.000  retained)
                         1.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
is_business_day? on weekend
                         0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
is_business_day? on weekday
                         0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
is_business_day? on holiday
                         0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(0) on weekend
                         0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(0) on weekday
                         0.000  memsize (     0.000  retained)
                         0.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(0) on holiday
                        72.000  memsize (     0.000  retained)
                         1.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(0) on weekend with monday holiday
                       144.000  memsize (     0.000  retained)
                         2.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(1) on weekend
                        72.000  memsize (     0.000  retained)
                         1.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(1) on weekday
                        72.000  memsize (     0.000  retained)
                         1.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(1) on holiday
                       144.000  memsize (     0.000  retained)
                         2.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(1) on weekend with monday holiday
                       216.000  memsize (     0.000  retained)
                         3.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(10) on weekend
                         1.008k memsize (     0.000  retained)
                        14.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(10) on weekday
                         1.008k memsize (     0.000  retained)
                        14.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(10) on holiday
                         1.080k memsize (     0.000  retained)
                        15.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
add_business_day(10) on weekend with monday holiday
                         1.152k memsize (     0.000  retained)
                        16.000  objects (     0.000  retained)
                         0.000  strings (     0.000  retained)
```

### Testing Script
```
$LOAD_PATH << '.'
$LOAD_PATH << 'lib'
$LOAD_PATH << 'lib/business_calendar'
require 'benchmark/ips'
require 'benchmark/memory'
require 'active_support'
require 'lib/business_calendar'
require 'active_support/core_ext/date'
require 'active_support/core_ext/numeric/time'
require 'active_support/core_ext/date/calculations'
class GCSuite
  def warming(*)
    run_gc
  end
  def running(*)
    run_gc
  end
  def warmup_stats(*)
  end
  def add_report(*)
  end
  private
  def run_gc
    GC.enable
    GC.start
    GC.disable
  end
end
suite = GCSuite.new
calendar = BusinessCalendar.for(:GB)
weekday = Date.today
weekend = Date.today - 2.days
holiday = Date.parse('2018-05-07')
weekend_before_holiday = holiday - 1.day
[ :ips, :memory ].each do |metric|
  Benchmark.send(metric) do |bm|
    bm.config(:time => 2, :warmup => 1, :suite => suite) if metric == :ips
    bm.report('day + 1') { weekday + 1 }
    bm.report('add_business_day(0) on weekend') { calendar.add_business_day(weekend, 0) }
    bm.report('add_business_day(0) on weekday') { calendar.add_business_day(weekday, 0) }
    bm.report('add_business_day(0) on holiday') { calendar.add_business_day(holiday, 0) }
    bm.report('add_business_day(0) on weekend with monday holiday') { calendar.add_business_day(weekend_before_holiday, 0) }
    bm.report('add_business_day(1) on weekend') { calendar.add_business_day(weekend, 1) }
    bm.report('add_business_day(1) on weekday') { calendar.add_business_day(weekday, 1) }
    bm.report('add_business_day(1) on holiday') { calendar.add_business_day(holiday, 1) }
    bm.report('add_business_day(1) on weekend with monday holiday') { calendar.add_business_day(weekend_before_holiday, 0) }
    bm.report('add_business_day(10) on weekend') { calendar.add_business_day(weekend, 10) }
    bm.report('add_business_day(10) on weekday') { calendar.add_business_day(weekday, 10) }
    bm.report('add_business_day(10) on holiday') { calendar.add_business_day(holiday, 10) }
    bm.report('add_business_day(10) on weekend with monday holiday') { calendar.add_business_day(weekend_before_holiday, 10) }
  end
end
```